+++ firmware/controllers/algo/advance_map.h
@@ -9,18 +9,18 @@
 
 #pragma once
 
-angle_t getWrappedAdvance(int rpm, float engineLoad);
-angle_t getCylinderIgnitionTrim(size_t cylinderNumber, int rpm, float ignitionLoad);
+angle_t getWrappedAdvance(float rpm, float engineLoad);
+angle_t getCylinderIgnitionTrim(size_t cylinderNumber, float rpm, float ignitionLoad);
 /**
  * this method is used to build default advance map
  */
-float getInitialAdvance(int rpm, float map, float advanceMax);
+float getInitialAdvance(float rpm, float map, float advanceMax);
 // public only for unit tests
-angle_t getCrankingAdvance(int rpm, float engineLoad);
-angle_t getRunningAdvance(int rpm, float engineLoad);
+angle_t getCrankingAdvance(float rpm, float engineLoad);
+angle_t getRunningAdvance(float rpm, float engineLoad);
 angle_t getAdvanceCorrections(float engineLoad);
 
-size_t getMultiSparkCount(int rpm);
+size_t getMultiSparkCount(float rpm);
 void initIgnitionAdvanceControl();
 
 class IgnitionState : public ignition_state_s {
